Added assets folder
Let's put the scripts, prefabs, scenes, and any other important assets in the folder.

Scene 2 is the most recent scene. It should contain a player, snail, and a pipe system for the snail to travel through.

I also included an exported Unity package file in case it's easier to export it in.